Dynamic Entity Management, Waypoints
- Added Dynamic Entity Management.  This is an automated system that
hides and reveals entities based on the location of online players.
Basically if an entity doesn't require to exist in the world because no
one would see it, it is hidden and removed so that it doesn't get
processed.  This optimization improves both FPS and UPS greatly in some
cases.
- Waypoint System.  This allows users to add personal waypoints that
only they can see that appear on their hud like beacons or antennas,
except it doesn't require blocks.  This will be expanded to allow
faction only waypoints as well.
- A whole lot of changes, bug fixes and updates.