Cleanup Triggers, Waypoint changes, Bug Fixes
- Added the new command /admin scan grids and /admin delete grids.  This
is a more sane scan / filter command that can be used with the new
Cleanup triggers introduced.
- Added cleanup triggers.  These are triggers based on /admin scan grids
that are trigger depending on a capacity.  So if your /admin scan grids
returns a number of ships greater than capacity, it will trigger a
cleanup X minutes after the trigger.  You will be able to warn users
about the cleanup as well before it happens.
- Added timed cleanups.  Based on the /admin scan grids command as well,
but runs at a set time in the day.
- Added /motd command which lets users see the login dialog again.
- Added /admin restart <time> to force a restart, and give warning.
- Added server waypoints.  These are waypoints that can not be removed
by a user.  This allows adding X, Y, Z style reference waypoints as well
as Center, without having to have physical stations at the locations.
- Added default waypoints.  These are waypoints added for users if they
have no waypoints.  Good for defining asteroid locations for new users.
- Fixed some issue with comm beacons being removed from the edges of the
world.