Docking Fixes, Small Additions, Code organization
- Fixed docking from using old API to ModAPI.  This will make it a bit
more reliable when it comes to docking and undocking.  Added /admin
undock all to force an undocking of ships that are docked.  I may add
more commands to allow fine control over undocking.  Docking is still
something slated to be completed (requires indexing on multiple ship
docks, and to not have to use LIFO to get a ship out of a dock)
- Added the ability to override the respawn menu.  This should only be
used on slow servers (that have a UPS below 20), as it may cause more
troubles than it fixes.
- Added /lastseen command which shows when a user last logged in
- Added block management.  This is not complete, so please avoid it for
now (I was forced to release before fully finishing due to the update).
I will add a more formal discussion of what this is.  (Basically
disables all blocks that aren't a certain block type if no one is
around)
- Started work on GameModes.  These are where I will experiment with
mini games.  Right now I'm working on an asteroid conquest mode.  This
mode isn't complete but the initial code is there.
- Updates to turret and entity management.
- A little bit of file reorganization.
- Lots of small fixes.